requirements.txt
torch
torchvision
streamlit
Pillow
pandas
numpy
matplotlib
seaborn
scikit-learn
plotly
kaggle
requests
tqdm
